READ THE TOP OF GAMEMANAGER
pretty sure everything is working mostly. Will almost def need minor
changes but for the time being it should be fine. Feel free to mess
around with something or change it if its necessary. Try to make sure
your not breaking someone else's code if you mess with a command in
commandHandler or GameManager.